# Speed up "all up" service command