--- a/identity-service/scripts/dev-server.sh
+++ b/identity-service/scripts/dev-server.sh
@@ -7,7 +7,6 @@ link_libs=true
 if [ "$link_libs" = true ]
 then
     cd ../audius-libs
-    npm run dev &
     npm link
     cd ../app
     npm link @audius/libs