--- a/service-commands/.gitignore
+++ b/service-commands/.gitignore
@@ -1,3 +1,4 @@
 output.log
 error.log
+libs.log
 local-storage