--- a/service-commands/scripts/run-libs.sh
+++ b/service-commands/scripts/run-libs.sh
@@ -1,0 +1,2 @@
+# Convenience script to make it easier to grep for exactly what service command runs
+npm run dev