--- a/service-commands/src/setup.js
+++ b/service-commands/src/setup.js
@@ -790,7 +790,6 @@ const allUp = async ({
   if (buildDataEthContracts) {
     await runInSequence(prereqs, options)
   }
-  await runInSequence([[Service.USER_REPLICA_SET_MANAGER, SetupCommand.UP]])
 
   if (parallel) {
     const runParallel = async (up, healthCheck, register) => {
@@ -850,6 +849,9 @@ const allUp = async ({
     console.log('Registering services'.info)
     await runInSequence(nodeRegisterCommands.flat())
   }
+
+  // URSM has to up after Creator Nodes are registered
+  await runInSequence([[Service.USER_REPLICA_SET_MANAGER, SetupCommand.UP]])
 
   const durationSeconds = Math.abs((Date.now() - start) / 1000)
   console.log(`All services brought up in ${durationSeconds}s`.happy)