… service-commands/src/commands/service-commands.json