### Description

<!-- What is the purpose of this PR? What is the current behavior? New behavior? Relevant links and/or information pertaining to PR? -->

-  Rewrites `A up` a fair bit. Explanation below, but tl;dr: Don't wait 10 seconds before trying health check, and bring all services up before running health check
- Updates `parallel` option in `A up` to be `--parallel`
- Don't run `npm run dev` for libs in the containers ❗ This means you'll need to run it on the host to get realtime changes to the services ❗ 
  - The reason for this is every time a new service is brought up, it builds libs and that generates a new file which triggers existing linked services to restart, thus increasing the healthcheck time. It also means multiple service restarts every time libs is changed.
  - Follow up here might be to spin up a container for libs that simply runs `npm run dev` and mounts the folder? Or some `tmux` background script? Basically need only _one_ libs watcher/recompiler
- Moves URSM to be initialized first (can I do that?) because of all the complaints from each service that it wasn't running and also it makes sense to me
- Adjusts the parallelism to run _everything_ after the prereqs in parallel, not just DNs in parallel then CNs in parallel

#### Before This Change
One service at a time, that service brought up, wait 10 seconds (for creator nodes), checked to see if it was healthy (which waits 10 seconds before checking), and finally registered. All of that had to be completed before moving to the next node.

| Test Run # | Command | Seconds | Minutes |
| ---|---|---|---|
| 1 | `A up` | 114.334s | ~19min |
| 2 | `A up parallel` |  880.56s | ~15 min |
| 3 | `A up` | 894.928s | ~15 min |
| 4 | `A up` | 830.546s | ~14 min |
| 5 | `A up` | 804.998s | ~13 min |

#### After This Change
Each service is brought up (either sequentially or at the same time). Once they're _all_ up, _then_ we check if they're healthy. Once they're healthy, finally we register them.

| Test Run # | Command | Seconds | Minutes |
| ---|---|---|---|
| 1 | `A up` | 547.392s | ~9 min |
| 2 | `A up --parallel` |  406.314s | ~7 min |
| 3 | `A up --parallel` | 435.905s | ~7 min |
| 4 | `A up` | 506.012s | ~8 min |
| 5 | `A up --parallel` | 404.409s | ~7 min|


### Tests

<!-- List the manual tests and repro instructions to verify that this PR works as anticipated. Include log analysis if possible. If this change impacts clients, make sure that you have tested the clients! -->

Need volunteers to test this to make sure I didn't break anything

### How will this change be monitored? Are there sufficient logs?

<!-- For features that are critical or could fail silently please describe the monitoring/alerting being added. -->


<!--
================ REMINDER: ================
If this PR touches a critical flow (such as Indexing, Uploads, Gateway or the Filesystem), make sure to add the `requires-special-attention` label.

** Add relevant labels as necessary. **
-->

## TODO:

- [x] Comment out `--parallel` as it doesn't work with `shellEnv` the way it's set up.
- [x] Add running `npm run dev` inside of libs to `A up` as a new service command so that libs watches and transpiles on changes. Run on the host machine in the background, store the pid, and bring it down with `A down`