@dmanjunath 2022-05-05
this is a great direction, i agree that we don't need to wait 90s for a health check before starting to bring another node up. i volunteer to be a guinea pig for this! only thing i'm a little hesitant of is running the npm run dev & outside the container, in the push for simplicity i want to reduce moving pieces cause it's another thing to remember / forget