@rickyrombo 2022-05-05
> this is a great direction, i agree that we don't need to wait 90s for a health check before starting to bring another node up. i volunteer to be a guinea pig for this! only thing i'm a little hesitant of is running the npm run dev & outside the container, in the push for simplicity i want to reduce moving pieces cause it's another thing to remember / forget

I have the same concern... I wonder if we can bring up a new container just for watching and rebuilding libs? The key is we only want _one_ watcher/rebuilder. But also could leave the `npm run dev` in there since it only adds another minute or two.