@rickyrombo 2022-05-11
Thanks to @raymondjacobson and @dmanjunath I ironed out all the kinks and this thing is ready for primetime!

We tested this by running `A up` and `A up --parallel` and then seeding with `A seed create-user` and `A seed upload-track`. My last four runs on three different boxes were successful

The times from my parallel runs went up ~1 min from before the bug fixes (which included making sure DN was up first, since re-selection in libs seems to have a bug):
292.362s
291.257s
315.3s

But they're still better than without this change:
586.997s
608.867s
603.94s
589.536s

All I need now is a sign-off!